iamshiuba/PySTAR202504/requirements.txt
flask==3.0.0
jinja2==3.1.5
werkzeug==3.0.6
gunicorn==22.0.0
Flask-Compress==1.14
Flask-Limiter==3.5.0
prometheus-flask-exporter==0.23.0
pytest==7.4.3
python-dotenv
flask-dotenv